+++ lib/package/audit/formatter/risk.rb
@@ -1,5 +1,5 @@
 require_relative './base'
-require_relative '../bash_color'
+require_relative '../util/bash_color'
 
 module Package
   module Audit
@@ -13,11 +13,11 @@ module Package
         def format
           case @risk_type
           when Enum::RiskType::HIGH
-            BashColor.red(Enum::RiskType::HIGH)
+            Util::BashColor.red(Enum::RiskType::HIGH)
           when Enum::RiskType::MEDIUM
-            BashColor.orange(Enum::RiskType::MEDIUM)
+            Util::BashColor.orange(Enum::RiskType::MEDIUM)
           when Enum::RiskType::LOW
-            BashColor.yellow(Enum::RiskType::LOW)
+            Util::BashColor.yellow(Enum::RiskType::LOW)
           else
             @risk_type.to_s
           end
