+++ lib/package/audit/formatter/vulnerability.rb
@@ -1,5 +1,5 @@
 require_relative './base'
-require_relative '../bash_color'
+require_relative '../util/bash_color'
 
 module Package
   module Audit
@@ -13,11 +13,11 @@ module Package
         def format
           case @vulnerability
           when Enum::VulnerabilityType::UNKNOWN, Enum::VulnerabilityType::CRITICAL, Enum::VulnerabilityType::HIGH
-            BashColor.red(@vulnerability)
+            Util::BashColor.red(@vulnerability)
           when Enum::VulnerabilityType::MEDIUM
-            BashColor.orange(@vulnerability)
+            Util::BashColor.orange(@vulnerability)
           when Enum::VulnerabilityType::LOW
-            BashColor.yellow(@vulnerability)
+            Util::BashColor.yellow(@vulnerability)
           else
             @vulnerability
           end
