+++ lib/package/audit/ruby/bundler_specs.rb
@@ -2,7 +2,7 @@ require_relative '../dependency'
 require_relative './gem_meta_data'
 require_relative './vulnerability_finder'
 
-require 'bundler'
+require 'bundler/lockfile_parser'
 
 module Package
   module Audit
