--- sig/package/audit/bash_color.rbs
@@ -1,13 +1,0 @@
-module Package
-  module Audit
-    module BashColor
-      def self.green: (String?) -> String
-
-      def self.orange: (String?) -> String
-
-      def self.red: (String?) -> String
-
-      def self.yellow: (String?) -> String
-    end
-  end
-end
