+++ sig/package/audit/util/bash_color.rbs
@@ -1,0 +1,15 @@
+module Package
+  module Audit
+    module Util
+      module BashColor
+        def self.green: (String?) -> String
+
+        def self.orange: (String?) -> String
+
+        def self.red: (String?) -> String
+
+        def self.yellow: (String?) -> String
+      end
+    end
+  end
+end
